+++ ChangeLog.md
@@ -1,11 +1,12 @@
 # Version 0.2.1
 
 - New features:
-	* Components can now be directly evaluated on a domain by simply invoking them as function, e.g.
+	* Components can now be directly evaluated on a domain by invoking them as function, e.g.
 	```
 	comp = GModelFit.Gaussian(1, 0, 1);
 	comp(Domain(-5:5))
 	```
+
 - Performance improvements:
 	* During a fit the `Model.maincomp` is temporarily set to the main component name.  This allow avoiding unnecessay invocations of `find_maincomp()` during model evaluation;